+++ frontend/src/router/index.js
@@ -1,6 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
 
-import Home from "../components/Home.vue";
+import Home from "../components/home/Home.vue";
 import Genre from "../components/webtoon/Genre.vue";
 import Publish from "../components/webtoon/publish.vue";
 import Kakao from "../components/webtoon/Kakao.vue";
